requirements.txt
Flask==2.3.2
google-cloud-storage>=2.18.2
tensorflow==2.12.0